# pkg/docs/swagger.yaml
definitions:
  api.CreateUpdateIngredientRequest:
    type: object
  api.HealthzResponse:
    properties:
      ok:
        type: boolean
    type: object
  api.IngredientCrudResponse:
    properties:
      isCreated:
        type: boolean
      isDeleted:
        type: boolean
      isUpdated:
        type: boolean
    type: object
  api.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
    type: object
  model.Ingredient:
    type: object
info:
  contact: {}
paths:
  /healthz:
    get:
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HealthzResponse'
      summary: healthz asserts that the server is running
      tags:
      - example
  /ingredient:
    post:
      description: creates a new ingredient to be used for recipes
      parameters:
      - description: Ingredient to be created
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateUpdateIngredientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IngredientCrudResponse'
      summary: creates a new ingredient to be used for recipes
      tags:
      - ingredient
  /ingredients:
    get:
      description: list ingredients. By default it list the first 10 ingredients.
        Maximum number of ingredients to list is 100
      parameters:
      - description: number of elements to list. max 100
        in: query
        name: numItems
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Ingredient'
            type: array
      summary: list ingredients
      tags:
      - ingredient
  /logged-user:
    get:
      description: get the user from the current sessions by looking into the cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.User'
      summary: get the user from the current sessions
      tags:
      - user
swagger: "2.0"